VirtualTryOn/requirements.txt
# Web Framework
flask
flask-cors

# Azure Services
azure-identity
azure-eventhub

# Utilities
requests
python-dotenv